# Distinguish slower, same and faster comparisons